docs: explain usage of LinkExtension
